refactor: shift editors to individual files for discoverability
