+++ apps/studio/src/features/editing-experience/components/form-builder/renderers/TipTapEditor/index.ts
@@ -1,0 +1,3 @@
+export * from "./TiptapTextEditor"
+export * from "./TiptapAccordionEditor"
+export * from "./TiptapCalloutEditor"
